Split TH and variable stuff into its own modules
